requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
resend>=2.0.0
openai>=1.0.0
